requirements.txt
aiogram<4.0
pydantic-settings
requests~=2.31.0
mysql-connector-python